Refactor common locator code into common location
